Revert "integrate Lazy into std layout"
This reverts commit 237a97760ad79a21ce0655b9f5adc0cc5b5cbc79.
