+++ src/libcore/lib.rs
@@ -240,8 +240,6 @@ pub mod char;
 pub mod ffi;
 #[cfg(not(test))] // See #65860
 pub mod iter;
-#[unstable(feature = "once_cell", issue = "68198")]
-pub mod lazy;
 pub mod option;
 pub mod panic;
 pub mod panicking;
