+++ src/libcore/tests/lib.rs
@@ -43,7 +43,6 @@
 #![feature(option_unwrap_none)]
 #![feature(peekable_next_if)]
 #![feature(partition_point)]
-#![feature(once_cell)]
 #![feature(unsafe_block_in_unsafe_fn)]
 #![deny(unsafe_op_in_unsafe_fn)]
 
@@ -63,7 +62,6 @@ mod fmt;
 mod hash;
 mod intrinsics;
 mod iter;
-mod lazy;
 mod manually_drop;
 mod mem;
 mod nonzero;
